// src/presentation/localizacion-padre/interfaces/ILocalizacionPrueba.ts
export interface ILocalizacionPrueba {
  idLocalizacionPrueba: number | null;
  estadoLocalizacionPrueba: string | null;
  nombreLocalizacionPrueba: string;
  fechaLocalizacionPrueba: string; // Usamos string para fechas en formularios
  isLocalizacionPrueba: boolean;
  enteroLocalizacionPrueba: number | null;
  decimalLocalizacionPrueba: number | null;
  descripcionLocalizacionPrueba: string;
  estadoLocalizacion: string | null;
}